Mostly functional directional switching
Mostly functional version of direction switch! It doesn't work well for some rotations, but for the most part it's killer. Rotational connections allow entities to see and walk through them, and alter the entity's frame of reference as they do. For the player, this should maintain the illusion of a connected space for well-connected portals.