UI Manager
Added a UI manager, which controls the creation, deletion, and settings of windows. The UI manager saves it's settings to a separate file, so the user can edit the layout and have it save between sessions independent of the game save file. These changes required std::string serialization support, so that was added as well.