Added Tracy profiling + resource dependencies
Added in new profiling tools! Profiling is great, and it turns out that BearLibTerminal is REALLY slow. Fullscreen mode eats 12 out of our 16 ms, which is, as they say, not good. This likely means a library switch at some point in the near future. In other improvements, the resource system now supports dynamic dependencies. If one resource tries to load another during it's pack, it will be logged as a dependency and check for rebuilds. The resource system will now skip repacking if the source file doesn't exist (for packed release builds). Version checks for save files now silently fail, as if the file didn't exist.