The Fontain of Pain
Endless pain. It took more hours than I care to admit to get freetype to link properly. Freetype is now in the project with the associated FontManager, which handles the freetype initialization and wraps the library up so the resource user doesn't have to touch it. RogueFonts store a font and can load in the glyph images as RogueImages - this is missing a lot of support that fonts will eventually need but it loads monochrome fonts well enough for now. As a test, the vulkan test object preloads a bunch of characters and then loads one as a texture.