Clean up streams and IO
Created a data backend type, that lets any stream set how it is capturing and storing data. PackedFileStream is now just a file stream with a file data backend, and JSON streams can now read/write directly to files. Switched up the IO types to be template specializations of the input enum, so the functions using them can directly be templated on that enum.