Monsters (Pt 1)
First pass of the monster and movement systems. These are still very much a WIP, but I'm trying to build out a really flexible system for designing multi-tile monsters than move in interesting ways. This patch sets up monsters a bit (they don't save yet), and adds the first walking and crushing movement types to demo the capabilities. The first pass of the body interface is in, with the first multi-tile bodies coming probably in the next update!